Change how Nodes are stored in Graphs
Graphs used a vector of a vector of nodes, and this makes my life too
complicated, which is bad, complaxity bad. I did a vector of vectors of
size_t that will take the IDs of the nodes but the nodes are stored in a
vector. This eases access when setting the relationships between nodes.

+++ src/node.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include "node.hpp"
 
-YAPP::Node::Node(   unsigned int newId,
+YAPP::Node::Node(   size_t newId,
                     float x,
                     float y,
                     NodeGeometry nodeType) {
